v3.0 core: Hermite grains, cached pan, grain Shape, and the AIR page
The critical review found the granular engine -- the thing the plugin is
named for -- was its least refined component. This is the fix, plus one
new page built as a single coherent idea rather than more toggles.

Granular engine
- readInterpolated: 2-point linear -> 4-point Catmull-Rom Hermite. Linear
  interpolation acts as a phase-dependent low-pass, so pitched-up grains
  lost air and pitched-down grains aliased. This is the single biggest
  audible quality upgrade available in the engine.
- Equal-power pan gains are resolved once at spawn (Grain::panL/panR)
  instead of std::cos/std::sin per grain per sample -- ~12M transcendental
  calls/sec at 128 grains, for a value that never changes during a grain.
- New "Shape" parameter (ParamId::grainShape, full mod ring, fills the
  empty TEXTURE grid cell): 0 = percussive attack, 0.5 = symmetric Hann,
  1 = swell. Applied as a rational bias p/(p+s(1-p)) on window progress,
  so there is no pow() in the render loop; the exponent is computed on
  the message thread in setGrainShape(). Removed a dead, never-wired
  "grainShape" *choice* param that survived the earlier dead-param sweep.

AIR page (tab 5) -- "squelchy and high-end dynamic toys for the top
frequency ranges, automatable, with phasers and delays to layer tonality
into a note"
- AirEngine: LR4 split at airCrossover; the low band always passes clean
  and is delayed by the exciter's oversampler latency (integer ring) so
  recombination stays phase-coherent; the top band runs
  Squelch (TPT SVF LP/BP/HP, res 0.5..14, cutoff rides the band envelope
  +/-4 oct) -> Exciter (reuses the oversampled Saturator; always runs so
  latency is constant, mix=0 when off) -> Dyn Shelf (IIR high shelf, gain
  = amount*12dB*(env-thr), coefficients recomputed only on change) ->
  Phaser -> feedback Delay. Blend is processed-top vs dry-top.
- 23 air* parameters, six-row layout matching MACHINES, tab index 5.
- 3 presets: Air - Acid Squelch / Silk Lift / Comb Tone.
- 5 tests: bypass bit-exact, low band preserved <5% with everything
  cranked, top band changed, mix=0 energy-transparent through the LR
  allpass, bounded over 200 blocks at res=1/fb=0.9/drive=1.

Also fixed while here (both pre-existing, both fired every block in
debug):
- MixEngine::processParallel added `output` to itself when the processor
  passed the live buffer as `dry` (the needsDry==false path). Guard the
  dry pass on &dry != &output -- by construction that path has ~0 dry
  weight, so no audible change; it just stops relying on addFrom
  tolerating aliasing.
- Three non-ASCII string literals passed to String(const char*) (an em
  dash in the lock tooltip x19 knobs, the AIR header, the share tooltip).

Saturator gains getLatencySamples(). musicalRange() handles grainShape
explicitly. Version 3.0.0 (3.0 not 2.0: the updater's isNewer() must see
it as newer than 2.4). 35/35 tests. Debug standalone now runs with zero
assertions.

+++ plugin/include/GrainFreeze/Modulation/MPEVoiceTracker.h
@@ -108,7 +108,7 @@ private:
     void writeVoice (size_t slot, juce::MPENote n, bool isNew)
     {
         const float semi = (float) (n.initialNote - root.load (std::memory_order_relaxed))
-                         + n.totalPitchbendInSemitones;
+                         + (float) n.totalPitchbendInSemitones;
         voices[slot].noteId.store ((int) n.noteID, std::memory_order_relaxed);
         voices[slot].semitones.store (semi, std::memory_order_relaxed);
         voices[slot].pressure .store (n.pressure.asUnsignedFloat(), std::memory_order_relaxed);
